% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomIUCMST.R
\name{binomIUCMST}
\alias{binomIUCMST}
\title{Binomial IUCMST}
\usage{
binomIUCMST(models, flavor = "B", ...)
}
\arguments{
\item{models}{Object with model information}

\item{flavor}{Flavor of penalty from \code{c("B","A","N")}}

\item{...}{additional paramters}
}
\description{
Binomial IUCMST
}
