% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcICs.R
\name{calcICs}
\alias{calcICs}
\title{Calculate information criteria}
\usage{
calcICs(models, flavor = "B")
}
\arguments{
\item{models}{Object with model information}

\item{flavor}{Flavor of penalty from \code{c("B","A","N")}}
}
\description{
Calculate information criteria
}
