% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcShat.R
\name{calcShat}
\alias{calcShat}
\title{Calculate variance-covariance matris}
\usage{
calcShat(indLR)
}
\arguments{
\item{indLR}{Object with likelihood ratio information}
}
\description{
Calculate variance-covariance matris
}
