% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcICs.R
\name{calcZ}
\alias{calcZ}
\title{Calculate Z scores}
\usage{
calcZ(
  models,
  Shat = calcShat(models),
  ICs = calcICs(models, flavor),
  flavor = "B",
  ...
)
}
\arguments{
\item{models}{Object with model information}

\item{Shat}{Matrix of variance and covariance estimates}

\item{ICs}{Calculated information criteria}

\item{flavor}{Flavor of penalty from \code{c("B","A","N")}}

\item{...}{additional parameters}
}
\description{
Calculate Z scores
}
