% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCcolors.R
\name{CCcolors}
\alias{CCcolors}
\alias{CCorigcolors}
\title{Collaborative Cross colors}
\source{
<https://csbio.unc.edu/CCstatus/index.py?run=AvailableLines.information>
}
\description{
A vector of 8 colors for use with the mouse
Collaborative Cross and Diversity Outbreds. Pulled from 'qtl2' package.
}
\details{
`CCorigcolors` are the original eight colors for the Collaborative Cross founder strains.
`CCcolors` are slightly modified.
}
\examples{
data(CCcolors)
data(CCorigcolors)
}
\keyword{datasets}
