% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_effect.R
\name{driver_effect}
\alias{driver_effect}
\title{Driver effect on target and mediator}
\usage{
driver_effect(out, driver_levels = LETTERS[1:8])
}
\arguments{
\item{driver_levels}{levels of driver}

\item{object}{data frame with effects}
}
\description{
Compare driver effect on mediator, target and target adjusted by mediator.
}
\examples{
data(Tmem68)
# Find mediators with significant effect
# Find and remove Tmem68 from mediators because it is target.
m <- match("Tmem68", Tmem68$annotation$symbol)
Tmem68$annotation[m,]
med_lod <- mediator_lod(mediator = Tmem68$mediator[,-m],
                        driver = Tmem68$qtl.geno,
                        annotation = Tmem68$annotation[-m,],
                        covar_med = NULL)
med_signif <- med_lod$lod >= 5
# Add info column.
med_lod$info <- paste("chr =", med_lod$chr)

med_test <- mediation_test(target = Tmem68$target,
                      mediator = Tmem68$mediator[, med_signif, drop = FALSE],
                      driver = Tmem68$qtl.geno,
                      annotation = med_lod,
                      covar_tar = Tmem68$covar,
                      method = "double-lod-diff")
(sum_med <- summary(med_test))

}
