% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDefault.R
\name{fitDefault}
\alias{fitDefault}
\title{Fit target relative to driver}
\usage{
fitDefault(driver, target, kinship = NULL, addcovar = NULL,
  intcovar = NULL, weights = NULL, ...)
}
\arguments{
\item{driver}{A matrix of drivers, individuals x drivers}

\item{target}{A numeric vector of target values}

\item{kinship}{Optional kinship matrix.}

\item{addcovar}{An optional matrix of additive covariates.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have `names`
for individual identifiers. Ignored if `kinship` is provided.#'}

\item{nullcovar}{An optional matrix of additional additive
covariates that are used under the null hypothesis (of no driver)
but not under the alternative (with driver).}
}
\value{
A list containing
* `LR` - The overall likelihood ratio.
* `indLR` - Vector of individual contributions to the likelihood ratio.
* `df` - Model degrees of freedom.
Currently `kinship` and `weights` are ignored.
}
\description{
Fit a model for a target and get detailed results
about estimated coefficients and individuals contributions to the LOD score.
}
\details{
If \code{kinship} is absent, regression is performed.
If \code{kinship} is provided, a linear mixed model is used, with a
random effect estimated under the null hypothesis of no driver,
and then taken as fixed and known with driver.
The default version ignores kinship. See \code{\link[qtl2]{fit1}}
for use of \code{kinship}.
}
