% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.snp.probs.R
\name{get.snp.probs}
\alias{get.snp.probs}
\title{Get the SNP probs nearest to the given chr and position.}
\usage{
get.snp.probs(chr, pos, genoprobs, map, query_fxn)
}
\arguments{
\item{chr}{character indicating the chromosome to use. This must match
one of the names of genoprobs and map.}

\item{pos}{numeric value in Mb indicating the position on the chromosome.}

\item{genoprobs}{list containing 3-dimensional arrays of allele probs, one 
for each chromosome. In qtl2 format.
Samples in rows, 8 founders in columns, markers in slices.}

\item{map}{list containing numeric vectors of marker positions in Mb.}

\item{query_fxn}{function for querying the CC_SNP data base. Obtain this
by calling qtl2:::create_variant_query_func().}
}
\value{
an n x 1 numeric matrix.
}
\description{
Get the SNP probs nearest to the given chr and position.
}
