% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ensembl.genes.R
\name{get_ensembl_genes}
\alias{get_ensembl_genes}
\title{Get the ensembl genes on a given set of chromosomes.}
\usage{
get_ensembl_genes(chr = c(1:19, "X", "Y", "MT"), build = 90)
}
\arguments{
\item{chr}{character vector of chromosomes from which to get genes.}

\item{build}{integer that is the Ensembl build to use.}
}
\value{
data.frame with the genes on the given chromosomes.
}
\description{
Get the ensembl genes on a given set of chromosomes.
}
\examples{
ens = get_ensembl_genes(chr = "19", build = 90)
}
