% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation.R
\name{ggplot_mediation}
\alias{ggplot_mediation}
\alias{autoplot.mediation}
\title{GG plot of conditioned LOD scores against genomic positions}
\usage{
ggplot_mediation(med, col = "firebrick4", cex = 2,
  ylab = "Conditioned LOD", ...)

\method{autoplot}{mediation}(x, ...)
}
\arguments{
\item{med}{mediation object}

\item{col}{color of points (default "firebrick4")}

\item{cex}{character expansion (default 2)}

\item{ylab}{Y axis label (default "Conditioned LOD")}

\item{...}{additional arguments to \code{\link[qtl2ggplot]{ggplot_scan1}}}
}
\description{
Plot LOD statistics calculated by \code{mediation.scan} against genomic positions
using ggplot2.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno)
ggplot2::autoplot(med)

}
\seealso{
\code{\link{identify.mediation}}, \code{\link{kplot}}
}
