% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_effect.R
\name{mediation_effect}
\alias{mediation_effect}
\alias{ggplot_mediation_effect}
\alias{autoplot.mediation_effect}
\alias{plot.mediation_effect}
\title{Driver effect on target and mediator}
\usage{
mediation_effect(
  test_object,
  id_name = "id",
  driver_levels = unique(unlist(test_object$driver_levels))
)

ggplot_mediation_effect(
  object,
  colors = intermediate::CCcolors,
  max_facet = 12,
  size_geom = 2,
  ...
)

\method{autoplot}{mediation_effect}(object, ...)

\method{plot}{mediation_effect}(x, ...)
}
\arguments{
\item{test_object}{object of class \code{mediation_test}}

\item{id_name}{name of identifier column}

\item{driver_levels}{levels of driver}

\item{object}{object of class \code{mediation_effect}}

\item{colors}{colors for plotting}

\item{max_facet}{maximum number of facets}

\item{size_geom}{geometry size for points and lines}

\item{...}{additional parameters}

\item{x}{object of class \code{mediation_effect}}
}
\description{
Compare driver effect on mediator, target and target adjusted by mediator.
}
\examples{
example("mediation_test")
out <- mediation_effect(med_test, "symbol")
ggplot_mediation_effect(out)

}
