% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_index.R
\name{mediation_index}
\alias{mediation_index}
\alias{plot.mediation_index}
\alias{autoplot.mediation_index}
\alias{ggplot_mediation_index}
\title{Test mediation across set of indexed drivers}
\usage{
mediation_index(
  target,
  mediator,
  driver = NULL,
  annotation = NULL,
  covar_tar = NULL,
  covar_med = NULL,
  driver_med = NULL,
  driver_index = colnames(mediator),
  facet_name = "chr",
  index_name = "pos",
  ...
)

\method{plot}{mediation_index}(x, ...)

\method{autoplot}{mediation_index}(x, ...)

ggplot_mediation_index(
  x,
  response = c("pvalue", "IC"),
  alpha = 0.5,
  pattern_name = "pattern",
  ...
)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{driver}{vector or matrix with driver values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{driver_med}{driver array for mediators}

\item{driver_index}{index to driver array}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}

\item{...}{additional parameters}

\item{x}{object of class \code{mediation_index}}

\item{response}{type of response from \code{c("pvalue","IC")}}

\item{alpha}{ggplot2 transparency parameter with default \code{0.5}}

\item{pattern_name}{name of pattern with default \code{"pattern"}}
}
\description{
Test mediation across set of indexed drivers
}
\examples{
data(Tmem68, package = "Tmem68")
 
target <- Tmem68$target

# Mediator is Tmem68.
m <- grep("Tmem68", Tmem68$annotation$symbol)
mediator <- Tmem68$mediator[,m]
annotation <- Tmem68$annotation[m,, drop = FALSE]

# Reduce to SNP for B6
driver_SNP <- cbind(B6 = Tmem68$driver[,2], rest = 1 - Tmem68$driver[,2])
driver_med <- list(allele = Tmem68$driver, SNP = driver_SNP)

med_index <- mediation_index(target = target,
                      mediator = mediator,
                      driver = NULL,
                      annotation = annotation,
                      covar_tar = Tmem68$covar,
                      covar_med = Tmem68$covar,
                      driver_med = driver_med,
                      driver_index = names(driver_med),
                      index_name = "probs")
summary(med_index)
ggplot2::autoplot(med_index)

}
