% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_joint.R
\name{mediation_joint}
\alias{mediation_joint}
\alias{plot.mediation_joint}
\alias{autoplot.mediation_joint}
\alias{ggplot_mediation_joint}
\title{Fit joint likelihood for `target` and `mediator` given `driver`.}
\usage{
mediation_joint(
  target,
  mediator,
  driver,
  annotation,
  covar_tar = NULL,
  covar_med = NULL,
  driver_med = NULL,
  intcovar = NULL,
  fitFunction = fitDefault,
  index_name = "pos",
  ...
)

\method{plot}{mediation_joint}(x, ...)

\method{autoplot}{mediation_joint}(x, ...)

ggplot_mediation_joint(
  x,
  lod = FALSE,
  xlab = index_name,
  ylab = ylab_name,
  ...
)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{A data frame with mediators' annotation with columns for `facet_name` and `index_name`}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{driver_med}{optional driver matrix for mediators}

\item{intcovar}{optional interactive covariates (assumed same for `mediator` and `target`)}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{index_name}{name of index column (default `pos`)}

\item{...}{additional parameters}

\item{x}{object of class \code{mediation_joint}}

\item{lod}{plot lod if \code{TRUE}}

\item{xlab}{horizontal label}

\item{ylab}{vertical label}
}
\value{
Data frame with `id` and `LR` as well as `annotation` columns.
}
\description{
Fit joint likelihood for `target` and `mediator` given `driver`.
}
\examples{
data(Tmem68)
 
target <- Tmem68$target

# Find mediators with significant effect
med_LR <- mediator_LR(mediator = Tmem68$mediator,
                        driver = Tmem68$driver,
                        annotation = Tmem68$annotation,
                        covar_med = Tmem68$covar)
med_signif <- med_LR$id[med_LR$LR >= 5 * log(10)]
# Add info column.
med_LR$info <- paste("chr =", med_LR$chr)
# Rename mediator LR column to not conflict with mediation test LR.
med_LR <- dplyr::rename(med_LR, mediatorLR = "LR")

med_joint <- mediation_joint(target = target,
                      mediator = Tmem68$mediator[, med_signif, drop = FALSE],
                      driver = Tmem68$driver,
                      annotation = med_LR,
                      covar_tar = Tmem68$covar,
                      covar_med = Tmem68$covar)
                      
ggplot_mediation_joint(med_joint)

}
