% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_scan.R, R/mediation_scan.R
\name{ggplot_mediation_scan}
\alias{ggplot_mediation_scan}
\alias{autoplot.mediation_scan}
\alias{plot.mediation_scan}
\alias{mediation_scan}
\alias{subset.mediation_scan}
\alias{summary.mediation_scan}
\title{Plot of conditioned LR scores against index}
\usage{
ggplot_mediation_scan(
  x,
  col = "firebrick4",
  size = 1,
  xlab = index_name,
  ylab = "Conditioned LR",
  col_target = "blue",
  gap = 25
)

\method{autoplot}{mediation_scan}(x, ...)

\method{plot}{mediation_scan}(x, ...)

mediation_scan(
  target,
  mediator,
  driver,
  annotation,
  covar = NULL,
  intcovar = NULL,
  method = c("double-LR-diff", "ignore", "LR-diff"),
  fitFunction = fitDefault,
  facet_name = "chr",
  index_name = "pos",
  verbose = TRUE,
  ...
)

\method{subset}{mediation_scan}(x, facets = NULL, chrs = NULL, ...)

\method{summary}{mediation_scan}(object, n = 10, minimal = FALSE, ...)
}
\arguments{
\item{x, object}{mediation object}

\item{col}{color of points (default "firebrick4")}

\item{size}{character size (default 2)}

\item{xlab, ylab}{X and Y axis label (default `index_name` and "Conditioned LR")}

\item{col_target}{color for target LR line}

\item{gap}{gap between facets (default `25`)}

\item{...}{additional parameters}

\item{target}{A numeric vector with gene/protein expression}

\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{driver}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{annotation}{A data frame with mediators' annotation with columns `facet_name` and `index_name`}

\item{covar}{A matrix with additive covariates}

\item{intcovar}{A matrix of covariate interacting with driver}

\item{method}{A method to handle missing cases}

\item{fitFunction}{function to fit models}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}

\item{verbose}{If TRUE display information about the progress}

\item{facets}{names of facets to subset}

\item{chrs}{chromosome names to subset}

\item{n}{maximum number of mediators to show (default 10)}

\item{minimal}{show only symbol, facet_name, index_name and LR if \code{TRUE}}
}
\description{
Plot LR statistics calculated by [mediation_scan()] against index
using ggplot2.
}
\details{
For a given QTL haplotype probabilities `driver`` and target `target`,
the function sequentially tries to add each column of `mediator` matrix as a covariate
and calculates LR statistic. The low LR value indicates `driver` and
`target` are conditionally independent given `mediator`,
i.e. `mediator` is a mediator of causal relationship from `driver` to `target`.
}
\examples{
data(Tmem68)

target <- Tmem68$target
m <- match("Tmem68", Tmem68$annotation$symbol)

med_scan <- mediation_scan(target = target,
                      mediator = Tmem68$mediator,
                      driver = Tmem68$driver,
                      annotation = Tmem68$annotation,
                      covar = Tmem68$covar)
                      
summary(med_scan)

ggplot_mediation_scan(med_scan)

}
