% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator_LR.R
\name{mediator_LR}
\alias{mediator_LR}
\title{Mediator Likelihood Ratio}
\usage{
mediator_LR(
  mediator,
  driver,
  annotation,
  covar_med = NULL,
  intcovar = NULL,
  annotation_names = c(facet = "chr", index = "pos", driver = NULL),
  verbose = TRUE,
  fitFunction = fitDefault,
  signif = 5 * log(10),
  ...
)
}
\arguments{
\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{driver}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{annotation}{A data frame with mediators' annotation with columns for facet and index}

\item{covar_med}{A matrix with additive covariates}

\item{intcovar}{A matrix of covariate interacting with driver}

\item{annotation_names}{names in annotation of columns for facet, index and, optionally, driver (default `c(facet = "chr", index = "pos", driver = NULL)`)}

\item{verbose}{If TRUE display information about the progress}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{signif}{value of LR to be significant (default `5*log(10)`)}

\item{...}{additional parameters}
}
\description{
For a given QTL haplotype probabilities \code{driver} and target \code{target},
the function sequentially tries to add each column \code{m} of \code{mediator} matrix as a covariate
and calculates LR statistic. A low LR value indicates \code{driver} and
\code{target} are conditionally independent given \code{m},
i.e. \code{m} is a mediator of causal relationship from \code{driver} to \code{target}.
}
\examples{
data(Tmem68)

med_LR <- mediator_LR(mediator = Tmem68$mediator,
                      driver = Tmem68$driver,
                      annotation = Tmem68$annotation,
                      covar_med = NULL)
ggplot2::autoplot(med_LR) +
  ggplot2::geom_hline(yintercept = 5, col = "blue")

}
