% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.chrlen.R
\docType{data}
\name{mouse.chrlen}
\alias{mouse.chrlen}
\title{Mouse chromosome lengths}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://github.com/churchill-lab/intermediate}{Churchill GitHub}
}
\usage{
data(mouse.chrlen)
}
\description{
Lengths in base pairs of the 22 autosomes for mouse.
}
\examples{
data(mouse.chrlen)
mouse.chrlen
}
\references{
Chick et al. (2016) 
(\href{https://dx.org/10.1038/nature18270}{Nature 534: 500-505})
}
\keyword{datasets}
