% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normIUCMST.R
\name{normIUCMST}
\alias{normIUCMST}
\title{Normal IUCMST}
\usage{
normIUCMST(models, Zscores = calcZ(models, ...), ...)
}
\arguments{
\item{models}{Object with model information}

\item{Zscores}{Calculated Z scores from \code{calcZ}}

\item{...}{additional paramters}
}
\description{
Normal IUCMST
}
