% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normJointIUCMST.R
\name{normJointIUCMST}
\alias{normJointIUCMST}
\title{Joint Normal IUCMST}
\usage{
normJointIUCMST(
  models,
  Zscores = calcZ(models, ...),
  Shat = calcShat(models),
  ...
)
}
\arguments{
\item{models}{Object with model information}

\item{Zscores}{Calculated Z scores from \code{calcZ}}

\item{Shat}{Matrix of variance and covariance estimates}

\item{...}{additional parameters}
}
\description{
Joint Normal IUCMST
}
