% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmem68.lod.R
\docType{data}
\name{Tmem68.lod}
\alias{Tmem68.lod}
\title{Liver protein expression LOD scores}
\format{
An numeric vector.
}
\source{
\href{https://github.com/churchill-lab/intermediate}{Churchill GitHub}
}
\usage{
data(Tmem68.lod)
}
\description{
Data from LOD profiles of protein expression mapping experiment.
}
\examples{
data(Tmem68.lod)
str(Tmem68.lod)
}
\references{
Chick et al. (2016) 
(\href{https://dx.org/10.1038/nature18270}{Nature 534: 500-505})
}
\keyword{datasets}
