% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmem_helper.R
\name{Tmem_helper}
\alias{Tmem_helper}
\title{Helper function to extract subset chromosomes from Tmem68}
\usage{
Tmem_helper(object, chr = NULL)
}
\arguments{
\item{object}{list object with target, mediator, annotation, covar and qtl.geno}

\item{chr}{chromosome(s) to select (all if \code{NULL})}
}
\description{
Helper function to extract subset chromosomes from Tmem68
}
