% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcIUCMST.R
\name{wilcIUCMST}
\alias{wilcIUCMST}
\title{Wilcoxon IUCMST}
\usage{
wilcIUCMST(models, flavor = "B", ...)
}
\arguments{
\item{models}{Object with model information}

\item{flavor}{Flavor of penalty from \code{c("B","A","N")}}

\item{...}{additional paramters}
}
\description{
Wilcoxon IUCMST
}
