% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_effect.R
\name{driver_effect}
\alias{driver_effect}
\title{Driver effect on target and mediator}
\usage{
driver_effect(out, driver_levels = LETTERS[1:8])
}
\arguments{
\item{driver_levels}{levels of driver}

\item{object}{data frame with effects}
}
\description{
Compare driver effect on mediator, target and target adjusted by mediator.
}
