% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDefault.R
\name{fitDefault}
\alias{fitDefault}
\title{Fit target relative to driver}
\usage{
fitDefault(
  driver,
  target,
  addcovar = NULL,
  intcovar = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{driver}{A matrix of drivers, individuals x drivers}

\item{target}{A numeric vector of target values}

\item{addcovar}{An optional matrix of additive covariates.}

\item{intcovar}{An optional matrix of interactive covariates.}

\item{weights}{An optional vector of positive weights for the
individuals. As with the other inputs, it must have `names`
for individual identifiers.}

\item{...}{additional parameters}
}
\value{
A list containing
* `LR` - The overall likelihood ratio.
* `indLR` - Vector of individual contributions to the likelihood ratio.
* `df` - Model degrees of freedom.
Currently `weights` is ignored.
}
\description{
Fit a model for a target and get detailed results
about estimated coefficients and individuals contributions to the LOD score.
}
\details{
Perform regression under the null hypothesis of no driver,
and then taken as fixed and known with driver.
}
