% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_scan.R
\name{ggplot_mediation_scan}
\alias{ggplot_mediation_scan}
\alias{autoplot.mediation_scan}
\title{GG plot of conditioned LOD scores against genomic positions}
\usage{
ggplot_mediation_scan(x, col = "firebrick4", cex = 1,
  ylab = "Conditioned LOD", col_target = "blue", ...)

\method{autoplot}{mediation_scan}(x, ...)
}
\arguments{
\item{x}{mediation object}

\item{col}{color of points (default "firebrick4")}

\item{cex}{character expansion (default 2)}

\item{ylab}{Y axis label (default "Conditioned LOD")}

\item{...}{additional arguments to \code{\link[qtl2ggplot]{ggplot_scan1}}}
}
\description{
Plot LOD statistics calculated by \code{mediation.scan} against genomic positions
using ggplot2.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno)
ggplot2::autoplot(med)

}
\seealso{
\code{\link{identify.mediation}}, \code{\link{kplot}}
}
