% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_scan.R
\name{ggplot_mediation_scan}
\alias{ggplot_mediation_scan}
\alias{autoplot.mediation_scan}
\alias{plot.mediation_scan}
\title{Plot of conditioned LOD scores against genomic positions}
\usage{
ggplot_mediation_scan(x, col = "firebrick4", cex = 1,
  ylab = "Conditioned LOD", col_target = "blue", gap = 25,
  facet_name = "chr")

\method{autoplot}{mediation_scan}(x, ...)

\method{plot}{mediation_scan}(x, ...)
}
\arguments{
\item{x}{mediation object}

\item{col}{color of points (default "firebrick4")}

\item{cex}{character expansion (default 2)}

\item{ylab}{Y axis label (default "Conditioned LOD")}

\item{col_target}{color for target LOD line}

\item{gap}{gap between facets (default `25`)}

\item{facet_name}{name of facet column (default `chr`)}
}
\description{
Plot LOD statistics calculated by [mediation_scan()] against genomic positions
using ggplot2.
}
\examples{
data(Tmem68)
# Find and remove Tmem68 from mediators because it is target.
m <- match("Tmem68", Tmem68$annotation$symbol)
Tmem68$annotation[m,]
med_scan <- mediation_scan(target = Tmem68$target,
                      mediator = Tmem68$mediator[,-m],
                      driver = Tmem68$qtl.geno,
                      annotation = Tmem68$annotation[-m,],
                      covar = Tmem68$covar,
                      method = "double-lod-diff")
ggplot2::autoplot(med_scan)
ggplot2::autoplot(subset(med_scan, "4")) +
  ggplot2::geom_vline(xintercept = Tmem68$annotation[m,"pos"], linetype = "dashed")

}
