% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_effect.R
\name{mediation_effect}
\alias{mediation_effect}
\alias{autoplot.mediation_effect}
\alias{plot.mediation_effect}
\title{Driver effect on target and mediator}
\usage{
mediation_effect(object, id_name = "id",
  driver_levels = unique(unlist(object$driver_levels)))

\method{autoplot}{mediation_effect}(object, ...)

\method{plot}{mediation_effect}(x, ...)
}
\arguments{
\item{object}{object from [mediation_test()]}

\item{id_name}{name of identifier column}

\item{driver_levels}{levels of driver}
}
\description{
Compare driver effect on mediator, target and target adjusted by mediator.
}
\examples{
example("mediation_test")
out <- mediation_effect(med_test, "symbol")
ggplot_mediation_effect(out)

}
