% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_index.R
\name{mediation_index}
\alias{mediation_index}
\title{Test mediation across set of indexed drivers}
\usage{
mediation_index(target, mediator, driver = NULL, annotation = NULL,
  covar_tar = NULL, covar_med = NULL, kinship = NULL, driver_med = NULL,
  driver_index = NULL, index_name = "pos", ...)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{driver_med}{driver array for mediators}

\item{driver_index}{index to driver array}

\item{...}{additional parameters}
}
\description{
Test mediation across set of indexed drivers
}
