% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_listof_mediation_qtl2.R,
%   R/mediation_qtl2.R
\name{plot.listof_mediation_qtl2}
\alias{plot.listof_mediation_qtl2}
\alias{autoplot.listof_mediation_qtl2}
\alias{ggplot_listof_mediation_qtl2}
\alias{mediation_qtl2}
\alias{plot.mediation_qtl2}
\alias{autoplot.mediation_qtl2}
\alias{ggplot_mediation_qtl2}
\title{Test mediation across set of indexed drivers for `qtl2` data}
\usage{
\method{plot}{listof_mediation_qtl2}(x, ...)

\method{autoplot}{listof_mediation_qtl2}(x, ...)

ggplot_listof_mediation_qtl2(
  x,
  plot_type = c("all", "causal", "reactive", "independent", "undecided"),
  minpvalue = 0.05,
  id_name = "mediator_id",
  ...
)

mediation_qtl2(
  target,
  mediator,
  annotation,
  covar_tar,
  covar_med,
  kinship,
  genoprobs,
  map,
  drop_lod = 1.5,
  min_lod = 3,
  query_variant,
  cores = 1,
  target_scan,
  ...
)

\method{plot}{mediation_qtl2}(x, ...)

\method{autoplot}{mediation_qtl2}(x, ...)

ggplot_mediation_qtl2(
  x,
  response = c("pvalue", "IC"),
  pattern_name = "pattern",
  ...
)
}
\arguments{
\item{...}{additional parameters for [mediation_index()]}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{map}{list of map positions}

\item{drop_lod}{drop in `LOD` (= `LR/log(10)`) to define index set}

\item{query_variant}{function to query variant database}

\item{cores}{number of cores to use}

\item{target_scan}{optional object from [qtl2::scan1snps()] for target (created if missing)}

\item{genoprob}{genoprob object of class [qtl2::calc_genoprob()]}
}
\value{
Object of class `mediation_qtl2`, which inherits from class [mediation_index()]
}
\description{
Test mediation across set of indexed drivers for `qtl2` data
}
