% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_scan.R
\name{mediation_scan}
\alias{mediation_scan}
\alias{subset.mediation_scan}
\title{Mediation Scan}
\usage{
mediation_scan(target, mediator, driver, annotation, covar = NULL,
  method = c("double-lod-diff", "ignore", "lod-diff", "lod-ratio"),
  verbose = TRUE)

\method{subset}{mediation_scan}(object, chrs = NULL, ...)
}
\arguments{
\item{target}{A numeric vector with gene/protein expression}

\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{driver}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{annotation}{A data frame with mediators' annotation, must include columns "chr" and "pos"}

\item{covar}{A matrix with additive covariates}

\item{method}{A method to handle missing cases}

\item{verbose}{If TRUE display information about the progress}
}
\description{
For a given QTL haplotype probabilities \code{driver} and target \code{target},
the function sequentially tries to add each column \code{m} of \code{mediator} matrix as a covariate
and calculates LOD statistic. The low LOD value indicates \code{driver} and
\code{target} are conditionally independent given \code{m},
i.e. \code{m} is a mediator of causal relationship from \code{driver} to \code{target}.
}
\examples{
data(Tmem68)
# Find and remove Tmem68 from mediators because it is target.
m <- match("Tmem68", Tmem68$annotation$symbol)
Tmem68$annotation[m,]
med_scan <- mediation_scan(target = Tmem68$target,
                      mediator = Tmem68$mediator[,-m],
                      driver = Tmem68$qtl.geno,
                      annotation = Tmem68$annotation[-m,],
                      covar = Tmem68$covar,
                      method = "double-lod-diff")
ggplot2::autoplot(med_scan)

}
\seealso{
\code{\link{plot.mediation}}, \code{\link{kplot}}
}
