% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_scan.R, R/mediation_scan.R
\name{ggplot_mediation_scan}
\alias{ggplot_mediation_scan}
\alias{autoplot.mediation_scan}
\alias{plot.mediation_scan}
\alias{mediation_scan}
\alias{subset.mediation_scan}
\title{Plot of conditioned LOD scores against index}
\usage{
ggplot_mediation_scan(
  x,
  col = "firebrick4",
  cex = 1,
  xlab = index_name,
  ylab = "Conditioned LOD",
  col_target = "blue",
  gap = 25
)

\method{autoplot}{mediation_scan}(x, ...)

\method{plot}{mediation_scan}(x, ...)

mediation_scan(
  target,
  mediator,
  driver,
  annotation,
  covar = NULL,
  intcovar = NULL,
  kinship = NULL,
  method = c("double-lod-diff", "ignore", "lod-diff"),
  fitFunction = fitDefault,
  facet_name = "chr",
  index_name = "pos",
  verbose = TRUE,
  ...
)

\method{subset}{mediation_scan}(x, facets = NULL, ...)
}
\arguments{
\item{x}{object of class \code{mediation_scan}}

\item{col}{color of points (default "firebrick4")}

\item{cex}{character expansion (default 2)}

\item{xlab, ylab}{X and Y axis label (default `index_name` and "Conditioned LOD")}

\item{col_target}{color for target LOD line}

\item{gap}{gap between facets (default `25`)}

\item{...}{additional parameters}

\item{target}{A numeric vector with gene/protein expression}

\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{driver}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{annotation}{A data frame with mediators' annotation with columns `facet_name` and `index_name`}

\item{covar}{A matrix with additive covariates}

\item{intcovar}{A matrix of covariate interacting with driver}

\item{kinship}{kinship object}

\item{method}{A method to handle missing cases}

\item{fitFunction}{function to fit models}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}

\item{verbose}{If TRUE display information about the progress}

\item{facets}{names of facets to subset}
}
\description{
Plot LOD statistics calculated by [mediation_scan()] against index
using ggplot2.
}
\details{
For a given QTL haplotype probabilities `driver`` and target `target`,
the function sequentially tries to add each column of `mediator` matrix as a covariate
and calculates LOD statistic. The low LOD value indicates `driver` and
`target` are conditionally independent given `mediator`,
i.e. `mediator` is a mediator of causal relationship from `driver` to `target`.
}
\examples{
data(Tmem68)
target <- Tmem68$target
m <- match("Tmem68", Tmem68$annotation$symbol)

med_scan <- mediation_scan(target = target,
                      mediator = Tmem68$mediator,
                      driver = Tmem68$qtl.geno,
                      annotation = Tmem68$annotation,
                      covar = Tmem68$covar,
                      method = "double-lod-diff")
ggplot2::autoplot(med_scan)
ggplot2::autoplot(subset(med_scan, "4")) +
  ggplot2::geom_vline(xintercept = Tmem68$annotation[m,"pos"], linetype = "dashed")

}
