% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_test.R, R/mediation_test.R
\name{plot.mediation_test}
\alias{plot.mediation_test}
\alias{autoplot.mediation_test}
\alias{ggplot_mediation_test}
\alias{mediation_test}
\alias{subset.mediation_test}
\alias{summary.mediation_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
\method{plot}{mediation_test}(x, ...)

\method{autoplot}{mediation_test}(x, ...)

ggplot_mediation_test(
  x,
  type = c("pos_LR", "pos_pvalue", "pvalue_LR", "alleles", "mediator"),
  main = params$target,
  maxPvalue = 0.1,
  local_only = FALSE,
  significant = TRUE,
  lod = FALSE,
  target_index = NULL,
  colors = ...
)

mediation_test(
  target,
  mediator,
  driver,
  annotation = NULL,
  covar_tar = NULL,
  covar_med = NULL,
  driver_med = NULL,
  intcovar = NULL,
  test = c("wilcoxon", "binomial", "joint", "normal"),
  fitFunction = fitDefault,
  facet_name = "chr",
  index_name = "pos",
  ...
)

\method{subset}{mediation_test}(x, not_type, ...)

\method{summary}{mediation_test}(object, ..., lod = FALSE)
}
\arguments{
\item{x}{object of class \code{mediation_test}}

\item{...}{additional parameters}

\item{type}{type of plot from \code{c("pos_LR","pos_pvalue","pvalue_LR","alleles","mediator")}}

\item{main}{title for plot}

\item{maxPvalue}{maximum p-value with default \code{0.1}}

\item{local_only}{local only if \code{TRUE} (default \code{FALSE})}

\item{significant}{whow signficant if \code{TRUE} (default)}

\item{lod}{show as LOD plot if \code{TRUE} (default)}

\item{target_index}{include vertical line at target if not \code{NULL}}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{A data frame with mediators' annotation with columns for `facet_name` and `index_name`}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{driver_med}{optional driver matrix for mediators}

\item{intcovar}{optional interactive covariates (assumed same for `mediator` and `target`)}

\item{test}{Type of CMST test.}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}

\item{not_type}{biotypes to not include}

\item{object}{object of class \code{mediation_test}}
}
\value{
List with elements:
- best best fit table
- test causal test results in table
- driver list of driver names for target and mediator(s)
- normF Frobenius norm if using both target and mediator drivers
- params list of parameter settings for use by summary and plot methods
}
\description{
Develop mediation models from driver, target and mediator
}
\examples{
data(Tmem68)
 
target <- Tmem68$target

# Find mediators with significant effect
med_LR <- mediator_LR(mediator = Tmem68$mediator,
                        driver = Tmem68$driver,
                        annotation = Tmem68$annotation,
                        covar_med = Tmem68$covar)
med_signif <- med_LR$id[med_LR$LR >= 5 * log(10)]
# Add info column.
med_LR$info <- paste("chr =", med_LR$chr)

med_test <- mediation_test(target = target,
                      mediator = Tmem68$mediator[, med_signif, drop = FALSE],
                      driver = Tmem68$driver,
                      annotation = med_LR,
                      covar_tar = Tmem68$covar,
                      covar_med = Tmem68$covar)
summary(med_test)

ggplot_mediation_test(med_test)

}
