% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_test.R, R/mediation_test.R
\name{plot.mediation_test}
\alias{plot.mediation_test}
\alias{autoplot.mediation_test}
\alias{ggplot_mediation_test}
\alias{mediation_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
\method{plot}{mediation_test}(x, ...)

\method{autoplot}{mediation_test}(x, ...)

ggplot_mediation_test(
  x,
  type = c("pos_lod", "pos_pvalue", "pvalue_lod", "alleles", "mediator"),
  main = params$target,
  maxPvalue = 0.1,
  local_only = FALSE,
  significant = TRUE,
  lod = TRUE,
  target_index = NULL,
  ...
)

mediation_test(
  target,
  mediator,
  driver,
  annotation = NULL,
  covar_tar = NULL,
  covar_med = NULL,
  kinship = NULL,
  driver_med = NULL,
  intcovar = NULL,
  test = c("wilcoxon", "binomial", "joint", "normal"),
  fitFunction = fitDefault,
  facet_name = "chr",
  index_name = "pos",
  ...
)
}
\arguments{
\item{...}{additional parameters}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{A data frame with mediators' annotation with columns for `facet_name` and `index_name`}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{driver_med}{optional driver matrix for mediators}

\item{intcovar}{optional interactive covariates (assumed same for `mediator` and `target`)}

\item{test}{Type of CMST test.}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}
}
\value{
List with elements:
- best best fit table
- test causal test results in table
- driver list of driver names for target and mediator(s)
- normF Frobenius norm if using both target and mediator drivers
- params list of parameter settings for use by summary and plot methods
}
\description{
Develop mediation models from driver, target and mediator
}
\examples{
data(Tmem68)
 
target <- Tmem68$target

# Reconstruct 8-allele genotype probabilities.
driver <- cbind(A = 1 - apply(Tmem68$qtl.geno, 1, sum), Tmem68$qtl.geno)
rownames(driver) <- rownames(Tmem68$qtl.geno)

# Find mediators with significant effect
med_lod <- mediator_lod(mediator = Tmem68$mediator,
                        driver = driver,
                        annotation = Tmem68$annotation,
                        covar_med = Tmem68$covar)
med_signif <- med_lod$id[med_lod$lod >= 5]
# Add info column.
med_lod$info <- paste("chr =", med_lod$chr)

med_test <- mediation_test(target = target,
                      mediator = Tmem68$mediator[, med_signif, drop = FALSE],
                      driver = driver,
                      annotation = med_lod,
                      covar_tar = Tmem68$covar,
                      covar_med = Tmem68$covar)
summary(med_test)
ggplot2::autoplot(med_test)

}
