% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediation_test.R, R/mediation_test.R
\name{plot.mediation_test}
\alias{plot.mediation_test}
\alias{autoplot.mediation_test}
\alias{ggplot_mediation_test}
\alias{mediation_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
\method{plot}{mediation_test}(x, ...)

\method{autoplot}{mediation_test}(x, ...)

ggplot_mediation_test(x, type = c("pos_lod", "pos_pvalue", "pvalue_lod",
  "alleles", "mediator"), main = params$target, maxPvalue = 0.1,
  local_only = FALSE, significant = TRUE, lod = TRUE, ...)

mediation_test(target, mediator, driver, annotation, covar_tar = NULL,
  covar_med = NULL, kinship = NULL, driver_med = NULL, intcovar = NULL,
  test = c("wilcoxon", "binomial", "joint", "normal"), pos = NULL,
  fitFunction = fitQtl2, data_type = c("phenotype", "expression"), ...)
}
\arguments{
\item{...}{additional parameters}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{driver_med}{optional driver matrix for mediators}

\item{intcovar}{optional interactive covariates (assumed same for `mediator`` and `target``)}

\item{test}{Type of CMST test.}

\item{pos}{Position of driver.}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{data_type}{Type of mediator data.}
}
\description{
Develop mediation models from driver, target and mediator
}
\examples{
data(Tmem68)
# Find mediators with significant effect
# Find and remove Tmem68 from mediators because it is target.
m <- match("Tmem68", Tmem68$annotation$symbol)
Tmem68$annotation[m,]
med_lod <- mediator_lod(mediator = Tmem68$mediator[,-m],
                        driver = Tmem68$qtl.geno,
                        annotation = Tmem68$annotation[-m,],
                        covar_med = NULL)
med_signif <- med_lod$lod >= 5
# Add info column.
med_lod$info <- paste("chr =", med_lod$chr)

med_test <- mediation_test(target = Tmem68$target,
                      mediator = Tmem68$mediator[, med_signif, drop = FALSE],
                      driver = Tmem68$qtl.geno,
                      annotation = med_lod,
                      covar_tar = Tmem68$covar,
                      method = "double-lod-diff")
summary(med_test)
ggplot2::autoplot(med_test)

}
