% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_triad.R
\name{mediation_triad}
\alias{mediation_triad}
\alias{ggplot_mediation_triad}
\title{Triad scatter plot for mediator and target}
\usage{
mediation_triad(target, mediator, driver, covar_tar = NULL,
  covar_med = NULL, kinship = NULL, fitFunction = fitQtl2, ...)

ggplot_mediation_triad(x, tname = "target", mname = "mediator",
  dname = "driver", centerline = NULL, fitline = FALSE,
  main = paste(tname, "by", mname, "and", dname), colors = qtl2::CCcolors,
  ...)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{driver}{vector or matrix with driver values}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{...}{additional arguments to internal routine}

\item{x}{object of class \code{mediation_triad}}

\item{tname}{target name (default \code{"target"})}

\item{mname}{mediator name (default \code{"mediator"})}

\item{dname}{driver name (default \code{"driver"})}

\item{centerline}{horizontal line at value (default = \code{NULL}); set to \code{NA} for no line or \code{NULL} for mean}

\item{fitline}{include fit line from coefficients in \code{x} if \code{TRUE}}

\item{main}{main title (defautl \code{tname})}

\item{colors}{named colors to use if \code{fitline} is \code{TRUE}}

\item{sdp}{SNP distribution pattern for plot colors}

\item{allele}{Driver has alleles if \code{TRUE}, otherwise allele pairs.}

\item{\dots}{additional parameters for plotting}
}
\description{
Triad plot. Currently relies on \code{sdp} to provide lines, but want to use
coefficients from model fit with \code{\link{mediation_test}} to get lines for
each column of driver. Note that the plot uses column \code{info} to provide
additional information, which here is the \code{chr} of mediator. The plot uses
the mediator position on its home chromosome, which is not really what is wanted.
See package \code{qtl2shiny} for a more elegant use.
}
\examples{
data(Tmem68)
target <- cbind(Tmem68 = Tmem68$target)
# Pick strongest mediator that is not target.
m <- match("Nnt", Tmem68$annotation$symbol)
mediator <- Tmem68$mediator[, m, drop = FALSE]
colnames(mediator) <- "Nnt"
# Reconstruct 8-allele genotype probabilities.
driver <- cbind(A = 1 - apply(Tmem68$qtl.geno, 1, sum), Tmem68$qtl.geno)
rownames(driver) <- rownames(Tmem68$qtl.geno)

med_triad <- mediation_triad(target = target,
                      mediator = mediator,
                      driver = driver,
                      covar_tar = Tmem68$covar,
                      sdp = 2)
ggplot2::autoplot(med_triad, tname = "Tmem68", mname = "Nnt")

}
