% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator_lod.R
\name{mediator_lod}
\alias{mediator_lod}
\title{Mediator LOD}
\usage{
mediator_lod(
  mediator,
  driver,
  annotation,
  covar_med = NULL,
  intcovar = NULL,
  kinship = NULL,
  facet_name = "chr",
  index_name = "pos",
  verbose = TRUE
)
}
\arguments{
\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{driver}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{annotation}{A data frame with mediators' annotation with columns for `facet_name` and `index_name`}

\item{covar_med}{A matrix with additive covariates}

\item{intcovar}{A matrix of covariate interacting with driver}

\item{kinship}{kinship object}

\item{facet_name}{name of facet column (default `chr`)}

\item{index_name}{name of index column (default `pos`)}

\item{verbose}{If TRUE display information about the progress}
}
\description{
For a given QTL haplotype probabilities \code{driver} and target \code{target},
the function sequentially tries to add each column \code{m} of \code{mediator} matrix as a covariate
and calculates LOD statistic. The low LOD value indicates \code{driver} and
\code{target} are conditionally independent given \code{m},
i.e. \code{m} is a mediator of causal relationship from \code{driver} to \code{target}.
}
\examples{
data(Tmem68, package = "Tmem68")
med_lod <- mediator_lod(mediator = Tmem68$mediator,
                      driver = Tmem68$qtl.geno,
                      annotation = Tmem68$annotation,
                      covar_med = NULL)
ggplot2::autoplot(med_lod) +
  ggplot2::geom_hline(yintercept = 5, col = "blue")

}
